Fix dim should be float in meshing

+++ se_core/include/se/algorithms/meshing.hpp
@@ -150,7 +150,7 @@ namespace meshing {
     return index;
   }
 
-  inline bool checkVertex(const Eigen::Vector3f v, const int dim){
+  inline bool checkVertex(const Eigen::Vector3f v, const float dim){
     return (v(0) <= 0 || v(1) <=0 || v(2) <= 0 || v(0) > dim || v(1) > dim || v(2) > dim);
   }
 
@@ -167,7 +167,7 @@ namespace algorithms {
       std::vector<se::VoxelBlock<FieldType>*> blocklist;
       std::mutex lck;
       const int size = volume.size();
-      const int dim = volume.dim();
+      const float dim = volume.dim();
       volume.getBlockList(blocklist, false);
       std::cout << "Blocklist size: " << blocklist.size() << std::endl;
